pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mea-analyzer"
version = "1.0.0"
description = "Read, clean, organize, plot and statistically analyze microelectrode arrays data from Axion Biosystems NeuralMetrics CSVs."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Vitor Bueno" }]

dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "matplotlib",
  "pyyaml",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]